Fix initialization of G_Temp and Q_Temp.
Rather than initialize the first h_SL columns of G_Temp and Q_Temp from
the previous frame and/or first column of the rest of it, initialize the
h_SL columns directly before RATE*t_E[0] in the TF grid. The spec would
have us do otherwise.

The effects of this can be seen on al_sbr_gh_32_1.

SPECBUG